Polished the players page by changing some CSS and added the tooltip for the navbar
