+++ web-app/src/components/PlayerCard/Index.js
@@ -1,58 +1,24 @@
-import axios from "axios";
-import { useState, useEffect } from "react";
-import "./Styles.css";
+import React from "react";
 
-require("dotenv").config();
+import "./styles.css";
 
-const PlayerCard = () => {
-  const [playersList, setPlayersList] = useState([]);
-  const [download, setDownload] = useState(false);
-
-  const fetchPlayers = () => {
-    console.log("Fetch");
-    try {
-      axios
-        .get(process.env.REACT_APP_LOCALHOST_URL)
-        .then((res) => {
-          let result = res.data;
-          setPlayersList(result);
-          setDownload(true);
-          console.log("res", result);
-          console.log("data", res.data);
-        })
-        .catch((error) => {
-          console.log(error);
-        });
-    } catch (error) {
-      console.log(error);
-    }
+const Index = ({ playerInfo }) => {
+  const { firstname, lastname, birthdate, picture, position } = {
+    ...playerInfo,
   };
-
-  useEffect(() => {
-    console.log("I'm here");
-    fetchPlayers();
-  }, []);
-
+  console.log(playerInfo);
+  console.log(firstname, lastname, birthdate, picture, position);
   return (
-    <div className="all-players">
-      <h1 className="playerCard-title">JOUEURS EN OBSERVATION :</h1>
-      <div className="list-players">
-        {playersList.map((player) => {
-          return (
-            <div className="player-card">
-              <p>
-                {player.firstname} {player.lastname}
-              </p>
-              <p>
-                Taille: {player.height}cm, Poids: {player.weight}KG
-              </p>
-              <p>{player.birthdate}</p>
-            </div>
-          );
-        })}
+    <div className="player-card-container">
+      <img alt="player staff" className="player-picture" src={`${picture}`} />
+      <div className="player-card-info">
+        <p className="player-card-text">{firstname}</p>
+        <p className="player-card-text">{lastname}</p>
+        <p className="player-card-text">{birthdate}</p>
+        <p className="player-card-text">{position}</p>
       </div>
     </div>
   );
 };
 
-export default PlayerCard;
+export default Index;
