--- web-app/src/components/PlayerCard/Styles.css
@@ -1,20 +1,0 @@
-.all-players {
-    margin-top: 30px;
-}
-
-.list-players {
-    display: flex;
-    justify-content: space-around;
-}
-
-.player-card {
-    color: #28df99;
-    background-color: white;
-    margin: 0 15px;
-    padding: 10px;
-}
-
-.playerCard-title {
-    color: white;
-    margin-bottom: 20px;
-}
