+++ web-app/src/components/PlayerCard/styles.css
@@ -1,0 +1,41 @@
+
+* {
+    margin: 0px;
+    padding: 0px;
+  }
+  
+  .player-card-container {
+    color: white;
+    display: flex;
+    flex-direction: row;
+    margin: 1em;
+  }
+  
+  .player-picture {
+    height: 110px;
+    width: 110px;
+    border: solid 5px #28df99;
+    border-radius: 100px;
+    position: relative;
+    right: -7px;
+    z-index: 10;
+  }
+  
+  .player-card-info {
+    background-color: grey;
+    display: flex;
+    flex-direction: column;
+    height: 120px;
+    position: relative;
+    top: 20%;
+    left: -45px;
+    border-radius: 35px;
+    width: 250px;
+    z-index: 8;
+    text-align: center;
+    justify-content: space-around;
+  }
+  
+  .player-card-text {
+    margin-left: 20%;
+  }
