--- web-app/src/pages/index.js
@@ -1,1 +1,0 @@
-// For ergonomic reason, Index.js contains all the others js files from the pages folder
