+++ web-app/src/pages/players/Index.js
@@ -1,9 +1,14 @@
-import Navbar from "../../components/navbar/index";
+import Navbar from "../../components/navbar/Index";
 import LateralBar from "../../components/LateralBar/Index";
 import PlayerCard from "../../components/PlayerCard/Index";
 
+import axios from "axios";
+import React, { useState, useEffect } from "react";
+import { Link } from "react-router-dom";
+
 //Material UI
 import { makeStyles } from "@material-ui/core/styles";
+
 // import IconButton from "@material-ui/core/IconButton";
 import AddIcon from "@material-ui/icons/Add";
 import FilterListIcon from "@material-ui/icons/FilterList";
@@ -18,7 +23,7 @@ import "./styles.css";
 const useStyles = makeStyles((theme) => ({
   buttons: {
     marginLeft: theme.spacing(2),
-    marginTop: theme.spacing(4),
+    marginTop: theme.spacing(0),
   },
   elementMR: {
     marginRight: theme.spacing(1),
@@ -27,20 +32,42 @@ const useStyles = makeStyles((theme) => ({
     marginLeft: theme.spacing(2),
   },
   elementMT: {
-    marginTop: theme.spacing(4),
+    marginTop: theme.spacing(0),
   },
 }));
 
 function Index() {
   const classes = useStyles();
+
+  const [listPlayer, setListPlayer] = useState([]);
+  const [download, setDownload] = useState(false);
+
+  const fetchPlayers = async () => {
+    await axios
+      .get("http://localhost:5000/api/players")
+      .then((res) => {
+        let result = res.data;
+        setListPlayer(result);
+        setDownload(true);
+        console.log(listPlayer, download);
+      })
+      .catch((error) => {
+        console.log(error);
+      });
+  };
+
+  useEffect(() => {
+    fetchPlayers();
+  }, []);
+
   return (
     <div className="player-page">
       <Navbar />
-      <div className="player-page-elements">
+      <div className="player-page-container">
         <LateralBar />
-        <div className="players-info">
+        <div className="player-page-content">
           <h1 className="player-title">JOUEURS</h1>
-          <div className="player-search-selection">
+          <div className="player-page-search-info">
             <TextField
               id="filled-basic"
               label="Search"
@@ -60,7 +87,15 @@ function Index() {
               </Fab>
             </Tooltip>
           </div>
-          <PlayerCard />
+          <div className="player-page-list">
+            {download ? (
+              listPlayer.map((data) => {
+                return <PlayerCard playerInfo={data} />;
+              })
+            ) : (
+              <div>No player in database</div>
+            )}
+          </div>
         </div>
       </div>
     </div>
