+++ web-app/src/pages/players/styles.css
@@ -2,20 +2,43 @@
     margin: 0px;
     padding: 0px;
   }
-
-.player-page{
+  
+  .player-page {
     background-color: black;
-}
-
-.player-title {
-    color: white;
-}
-
-.player-page-elements {
+  }
+  
+  .player-page-container {
     display: flex;
     flex-direction: row;
-}
-
-.players-info {
-    margin: 30px 0 0 50px; /*margin top 30px, margin left 10%*/
-}
+  }
+  
+  .player-page-content {
+    display: flex;
+    flex-direction: column;
+    width: 94%;
+  }
+  
+  .player-title {
+    color: white;
+    font-size: 3em;
+    margin: 3%;
+  }
+  
+  .player-page-search-info {
+    display: flex;
+    flex-direction: row;
+    margin-left: 3%;
+    max-width: 40%;
+  }
+  
+  .player-page-list {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    margin-left: 10%;
+    margin-top: 3%;
+    color: white;
+    width: 100%;
+    height: 65%;
+    overflow: auto;
+  }
